+++ rerun_py/rerun_sdk/rerun/__init__.py
@@ -125,44 +125,6 @@ __all__ = [
 ]
 
 
-# TODO(jleibs): docstrings
-def add_space_view(
-    name: str,
-    space_path: str,
-    entity_paths: List[str],
-    blueprint: Optional[RecordingStream] = None,
-) -> None:
-    blueprint = RecordingStream.to_native(blueprint)
-    bindings.add_space_view(name, space_path, entity_paths, blueprint)
-
-
-# TODO(jleibs): docstrings
-def set_panels(
-    *,
-    all_expanded: Optional[bool] = None,
-    blueprint_view_expanded: Optional[bool] = None,
-    selection_view_expanded: Optional[bool] = None,
-    timeline_view_expanded: Optional[bool] = None,
-    blueprint: Optional[RecordingStream] = None,
-) -> None:
-    blueprint = RecordingStream.to_native(blueprint)
-    bindings.set_panels(
-        blueprint_view_expanded=blueprint_view_expanded or all_expanded,
-        selection_view_expanded=selection_view_expanded or all_expanded,
-        timeline_view_expanded=timeline_view_expanded or all_expanded,
-        blueprint=blueprint,
-    )
-
-
-# TODO(jleibs): docstrings
-def set_auto_space_views(
-    enabled: bool,
-    blueprint: Optional[RecordingStream] = None,
-) -> None:
-    blueprint = RecordingStream.to_native(blueprint)
-    bindings.set_auto_space_views(enabled, blueprint)
-
-
 # If `True`, we raise exceptions on use error (wrong parameter types etc).
 # If `False` we catch all errors and log a warning instead.
 _strict_mode = False
@@ -176,10 +138,10 @@ def init(
     recording_id: Optional[str] = None,
     spawn: bool = False,
     init_logging: bool = True,
-    init_blueprint: bool = True,
-    add_to_app_default_blueprint: bool = True,
     default_enabled: bool = True,
     strict: bool = False,
+    exp_init_blueprint: bool = False,
+    exp_add_to_app_default_blueprint: bool = True,
 ) -> None:
     """
     Initialize the Rerun SDK with a user-chosen application id (name).
@@ -219,13 +181,13 @@ def init(
         Can overridden with the RERUN env-var, e.g. `RERUN=on` or `RERUN=off`.
     init_logging
         Should we initialize the logging for this application?
-    init_blueprint
-        Should we initialize the blueprint for this application?
-    add_to_app_default_blueprint
-        Should the blueprint append to the existing app-default blueprint instead of creating a new one.
     strict
         If `True`, an exceptions is raised on use error (wrong parameter types etc).
         If `False`, errors are logged as warnings instead.
+    exp_init_blueprint
+        (Experimental) Should we initialize the blueprint for this application?
+    exp_add_to_app_default_blueprint
+        (Experimental) Should the blueprint append to the existing app-default blueprint instead of creating a new one.
 
     """
 
@@ -240,14 +202,14 @@ def init(
             spawn=False,
             default_enabled=default_enabled,
         )
-    if init_blueprint:
-        new_blueprint(
+    if exp_init_blueprint:
+        experimental.new_blueprint(
             application_id=application_id,
             blueprint_id=recording_id,
             make_default=True,
             make_thread_default=False,
             spawn=False,
-            add_to_app_default_blueprint=add_to_app_default_blueprint,
+            add_to_app_default_blueprint=exp_add_to_app_default_blueprint,
             default_enabled=default_enabled,
         )
 
@@ -357,84 +319,6 @@ def new_recording(
         _spawn(recording=recording)
 
     return recording
-
-
-def new_blueprint(
-    application_id: str,
-    blueprint_id: Optional[str] = None,
-    make_default: bool = False,
-    make_thread_default: bool = False,
-    spawn: bool = False,
-    add_to_app_default_blueprint: bool = False,
-    default_enabled: bool = True,
-) -> RecordingStream:
-    """
-    Creates a new blueprint with a user-chosen application id (name) to configure the appearance of Rerun.
-
-    If you only need a single global blueprint, [`rerun.init`][] might be simpler.
-
-    Parameters
-    ----------
-    application_id : str
-        Your Rerun recordings will be categorized by this application id, so
-        try to pick a unique one for each application that uses the Rerun SDK.
-
-        For example, if you have one application doing object detection
-        and another doing camera calibration, you could have
-        `rerun.init("object_detector")` and `rerun.init("calibrator")`.
-    blueprint_id : Optional[str]
-        Set the blueprint ID that this process is logging to, as a UUIDv4.
-
-        The default blueprint_id is based on `multiprocessing.current_process().authkey`
-        which means that all processes spawned with `multiprocessing`
-        will have the same default recording_id.
-
-        If you are not using `multiprocessing` and still want several different Python
-        processes to log to the same Rerun instance (and be part of the same blueprint),
-        you will need to manually assign them all the same blueprint_id.
-        Any random UUIDv4 will work, or copy the recording id for the parent process.
-    make_default : bool
-        If true (_not_ the default), the newly initialized blueprint will replace the current
-        active one (if any) in the global scope.
-    make_thread_default : bool
-        If true (_not_ the default), the newly initialized blueprint will replace the current
-        active one (if any) in the thread-local scope.
-    spawn : bool
-        Spawn a Rerun Viewer and stream logging data to it.
-        Short for calling `spawn` separately.
-        If you don't call this, log events will be buffered indefinitely until
-        you call either `connect`, `show`, or `save`
-    add_to_app_default_blueprint
-        Should the blueprint append to the existing app-default blueprint instead instead of creating a new one.
-    default_enabled
-        Should Rerun logging be on by default?
-        Can overridden with the RERUN env-var, e.g. `RERUN=on` or `RERUN=off`.
-
-    Returns
-    -------
-    RecordingStream
-        A handle to the [`rerun.RecordingStream`][]. Use it to log data to Rerun.
-
-    """
-
-    blueprint_id = application_id if add_to_app_default_blueprint else blueprint_id
-
-    blueprint = RecordingStream(
-        bindings.new_blueprint(
-            application_id=application_id,
-            blueprint_id=blueprint_id,
-            make_default=make_default,
-            make_thread_default=make_thread_default,
-            default_enabled=default_enabled,
-        )
-    )
-
-    if spawn:
-        from rerun.sinks import spawn as _spawn
-
-        _spawn(recording=blueprint)
-
-    return blueprint
 
 
 def version() -> str:
