+++ rerun_py/rerun_sdk/rerun/experimental.py
@@ -5,8 +5,7 @@ These features are not yet stable and may change in future releases without
 going through the normal deprecation cycle.
 """
 
+from rerun.log.experimental.blueprint import add_space_view, new_blueprint, set_auto_space_views, set_panels
 from rerun.log.experimental.text import log_text_box
 
-__all__ = [
-    "log_text_box",
-]
+__all__ = ["log_text_box", "add_space_view", "set_panels", "set_auto_space_views", "new_blueprint"]
